+++ torch/utils/tensorboard/__init__.py
@@ -1,6 +1,6 @@
 """A module for visualization with tensorboard
 """
 
-from tensorboard.compat.tensorboardX.record_writer import RecordWriter
+from tensorboard.compat.tensorboard.record_writer import RecordWriter
 from .torchvis import TorchVis
 from .writer import FileWriter, SummaryWriter
