+++ torch/utils/tensorboard/beholder/beholder.py
