+++ torch/utils/tensorboard/embedding.py
