+++ torch/utils/tensorboard/utils.py
