+++ torch/utils/tensorboard/visdom_writer.py
